Move Repository interface to domain module.
