+++ data/repository/src/main/java/jp/kuaddo/tsuidezake/data/repository/internal/di/RepositoryModule.kt
@@ -2,8 +2,8 @@ package jp.kuaddo.tsuidezake.data.repository.internal.di
 
 import dagger.Binds
 import dagger.Module
-import jp.kuaddo.tsuidezake.data.repository.Repository
 import jp.kuaddo.tsuidezake.data.repository.internal.RepositoryImpl
+import jp.kuaddo.tsuidezake.domain.Repository
 
 @Module
 internal abstract class RepositoryModule {
