+++ domain/build.gradle.kts
@@ -10,7 +10,6 @@ apply<CommonBuildPlugin>()
 
 dependencies {
     implementation(project(":model"))
-    implementation(project(":data:repository"))
 
     implementation(Deps.Kotlin.stdlib)
     implementation(Deps.Kotlin.coroutinesCore)
