+++ domain/src/main/java/jp/kuaddo/tsuidezake/domain/GetSakeDetailUseCase.kt
@@ -1,6 +1,5 @@
 package jp.kuaddo.tsuidezake.domain
 
-import jp.kuaddo.tsuidezake.data.repository.Repository
 import jp.kuaddo.tsuidezake.model.Resource
 import jp.kuaddo.tsuidezake.model.SakeDetail
 import javax.inject.Inject
