+++ domain/src/main/java/jp/kuaddo/tsuidezake/domain/RemoveSakeFromTastedListUseCase.kt
@@ -1,6 +1,5 @@
 package jp.kuaddo.tsuidezake.domain
 
-import jp.kuaddo.tsuidezake.data.repository.Repository
 import jp.kuaddo.tsuidezake.model.Resource
 import javax.inject.Inject
